+++ test/Response/SapiEmitterTest.php
@@ -10,9 +10,10 @@
 namespace ZendTest\Diactoros\Response;
 
 use PHPUnit_Framework_TestCase as TestCase;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamInterface;
 use Zend\Diactoros\Response;
 use Zend\Diactoros\Response\SapiEmitter;
-use Zend\Diactoros\Stream;
 use ZendTest\Diactoros\TestAsset\HeaderStack;
 
 class SapiEmitterTest extends TestCase
@@ -45,6 +46,7 @@ class SapiEmitterTest extends TestCase
         ob_end_clean();
         $this->assertContains('HTTP/1.1 200 OK', HeaderStack::stack());
         $this->assertContains('Content-Type: text/plain', HeaderStack::stack());
+        $this->assertContains('Content-Length: 8', HeaderStack::stack());
     }
 
     public function testEmitsMessageBody()
@@ -56,5 +58,22 @@ class SapiEmitterTest extends TestCase
 
         $this->expectOutputString('Content!');
         $this->emitter->emit($response);
+    }
+
+    public function testDoesNotInjectContentLengthHeaderIfStreamSizeIsUnknown()
+    {
+        $stream = $this->prophesize('Psr\Http\Message\StreamInterface');
+        $stream->__toString()->willReturn('Content!');
+        $stream->getSize()->willReturn(null);
+        $response = (new Response())
+            ->withStatus(200)
+            ->withBody($stream->reveal());
+
+        ob_start();
+        $this->emitter->emit($response);
+        ob_end_clean();
+        foreach (HeaderStack::stack() as $header) {
+            $this->assertNotContains('Content-Length:', $header);
+        }
     }
 }